Allocate M5.31: rename the curvature scale instrument, add task record + roadmap row
