+++ openwave/xperiments/m5_liquid_crystal/research/m5_roadmap.md
@@ -100,6 +100,7 @@
 | [M5.21.11](tasks/m5_21_11_task_details.md) | The realistic-parameter bridge: the mass-ratio read | ❌ Closed TERMINAL + APPROVED 2026-08-07 (12:10 go-to-review), audit 5C/1PARTIAL/0R. The frozen route-(b) ladder ran verbatim (42 rungs, ~50 CPU-h): F4 fired (boost-dressing gain FLAT in g, measured on endpoints AND retro-read in the M5.21.8 family) and F3 fired (usable rungs 1/0/0 vs the 6-floor). Route (b) closed; [M8.6](../../m8_mit/research/m8_roadmap.md) permanently gated. [Record](tasks/m5_21_11_task_details.md) | 2026-08-07 |
 | [M5.21.14](tasks/m5_21_14_task_details.md) | The (1/g) dressing term: symbolic 4×4 boost hedgehog | ✅ Closed COMPLETE + APPROVED 2026-08-09 (01:28 go-to-review), audit 7C/2P/0R. The term DERIVED + triple-verified; **the author's criterion MET at the term level** (T1_kin negative-semidefinite, ω² sector); the kin flip BULK, at threshold on the certified box; the flat gain and the m\*-critique both retrodicted; T1 alone UNBOUNDED below so the 3×3 handoff carries a mandatory guard. [Note](findings/m5_21_14_note.md) | 2026-08-09 |
 | [M5.21.15](tasks/m5_21_15_task_details.md) | The coupled ω-minimum: electron analog with angular momentum | ✅ Closed COMPLETE + APPROVED 2026-08-11 (02:23 go-to-review), 2 audit rounds, 0 structural refutations. **FREE minimization can NEVER give an interior ω-minimum** (the envelope-concavity theorem, measured on every channel and block); **WITH fixed J the minimum EXISTS at POSITIVE energy** (ω = 0.59, E = +115.9, narrow guard), its energy sign flipping with guard width in [0.02, 0.05] ([Q25](m5_question_tracker.md#q25-detail)). [Note](findings/m5_21_15_note.md) | 2026-08-11 |
+| [M5.31](tasks/m5_31_task_details.md) | The M5/Faber curvature scale curve: dense C(rho) instrument | ✅ Closed COMPLETE 2026-08-14, **first external contribution on the running-coupling debt path** ([@vantasnerdan](https://github.com/vantasnerdan), [PR #437](https://github.com/openwave-labs/openwave/pull/437)). The five-shell onset resolved into a dense 25-radius C(rho) curve on the M5.6.4b hedgehog, nonperiodic stencils, 11+16 gates, independently audited; **explicitly a form factor, NOT a renormalized coupling** (the C→g_R convention stays author-gated). [Note](findings/m5_31_coupling_curvature_note.md) | 2026-08-14 |
 
 ## CHANGE-LOG
