+++ openwave/xperiments/m5_liquid_crystal/research/plots/m5_31_coupling_curvature_scan.png
