+++ openwave/xperiments/m5_liquid_crystal/research/scripts/m5_31_coupling_curvature_field.py
